chore(clippy): clean 5 legacy lints + flip CI advisory -> mandatory
The 5 lints the May-12 CI sweep parked as advisory are now addressed:

mainline-service:
- billing/tap.rs: drop redundant `u128 -> u128` cast on receipt.value
- grpc/server.rs: return Result<_, Box<Status>> from
  build_attestation_and_cursor() — the Status variant was 176 bytes;
  callers updated to deref-or-unbox on Err.
- tests/grpc_end_to_end.rs: factor the 6-tuple ParsedAttestation into
  a type alias.

mainline-gateway:
- gateway.rs: replace `candidates.iter().cloned()` with
  `candidates.into_iter()` — the binding isn't used afterward.
- quorum.rs: introduce `pub type PayloadHashBucket = (Option<[u8;32]>,
  Vec<[u8;20]>)` and use it in both call sites.

.github/workflows/ci.yml: clippy is now blocking — `continue-on-error`
removed. All 91 workspace tests still pass.

+++ mainline-service/src/billing/tap.rs
@@ -341,7 +341,7 @@ impl<L: AllocationLookup> ReceiptVerifier for EscrowVerifier<L> {
         if signer != info.payer {
             return Err(TapError::InvalidSignature);
         }
-        if (receipt.value as u128) > info.escrow_available {
+        if receipt.value > info.escrow_available {
             return Err(TapError::InsufficientEscrow);
         }
